espp_payoff/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
Django==4.1.4
djangorestframework==3.14.0
importlib-metadata==5.1.0
Markdown==3.4.1
numpy==1.23.5
polygon-api-client==1.5.0
pytz==2022.6
scipy==1.9.3
sqlparse==0.4.3
urllib3==1.26.13
websockets==10.4
zipp==3.11.0
